+++ modules/fit/index.ts
@@ -5,6 +5,6 @@ export type {
 } from './types';
 export {
   BLOCKER_LABEL, BLOCKER_SHORT, CERTAINTY_LABEL, CERTAINTY_WEIGHT, DECISION_LABEL, FAMILIARITY_LABEL,
-  FIRM_CLASS_LABEL, GRADE_LABEL, GRADE_SCORE, LINK_LABEL, SENTIMENT_LABEL,
+  FIRM_CLASS_LABEL, GRADE_LABEL, GRADE_SIGN, GRADE_VERDICT, GRADE_SCORE, LINK_LABEL, SENTIMENT_LABEL,
 } from './types';
 export { assessmentFor, assessmentsForEntity, firmProfile, listAssessments, listFirmProfiles } from './repo';